setup

+++ day19/src/data.rs
@@ -1,8 +1,7 @@
-use regex::Regex;
 use derive_new::new;
+use regex::Regex;
 
-pub type AnswerDtype = i64;
-
+pub type AnswerDtype = u32;
 
 const PATTERN: &str = r"Blueprint (\d+): Each ore robot costs (\d+) ore. Each clay robot costs (\d+) ore. Each obsidian robot costs (\d+) ore and (\d+) clay. Each geode robot costs (\d+) ore and (\d+) obsidian.";
 
@@ -13,7 +12,6 @@ pub struct Money {
     pub obsidian: u32,
 }
 
-
 #[derive(new, Debug, Copy, Clone)]
 pub struct Blueprint {
     pub number: u32,
@@ -22,8 +20,6 @@ pub struct Blueprint {
     pub obsidian_cost: Money,
     pub geode_cost: Money,
 }
-
-
 
 pub fn import_data(data: &str) -> Vec<Blueprint> {
     let re = Regex::new(PATTERN).unwrap();
@@ -36,7 +32,10 @@ fn parse_str_to_num(s: &str) -> u32 {
 }
 
 pub fn parse(line: &str, re: &Regex) -> Blueprint {
-    let (_, [blueprint_num, ore_ore, clay_ore, obs_ore, obs_clay, geo_ore, geo_obs]) = re.captures(line).expect("capture with regex failed").extract();
+    let (_, [blueprint_num, ore_ore, clay_ore, obs_ore, obs_clay, geo_ore, geo_obs]) = re
+        .captures(line)
+        .expect("capture with regex failed")
+        .extract();
 
     Blueprint::new(
         parse_str_to_num(blueprint_num),
@@ -45,7 +44,6 @@ pub fn parse(line: &str, re: &Regex) -> Blueprint {
         Money::new(parse_str_to_num(obs_ore), parse_str_to_num(obs_clay), 0),
         Money::new(parse_str_to_num(geo_ore), 0, parse_str_to_num(geo_obs)),
     )
-
 }
 
 pub const TEST_DATA_1: &str = r#"Blueprint 1: Each ore robot costs 4 ore. Each clay robot costs 2 ore. Each obsidian robot costs 3 ore and 14 clay. Each geode robot costs 2 ore and 7 obsidian.
